% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERSUADE_output_functions.R
\name{f_generate_report}
\alias{f_generate_report}
\title{Generate PDF Report for a PERSUADE Analysis}
\usage{
f_generate_report(PERSUADE)
}
\arguments{
\item{PERSUADE}{A PERSUADE object returned by \code{\link[=f_PERSUADE]{f_PERSUADE()}}.}
}
\value{
A length-1 character string giving the absolute path to the generated
PDF, returned invisibly.
}
\description{
Save the PERSUADE object and render a PDF report using the bundled
\code{PERSUADE_output.Rmd} template.
}
\details{
The R Markdown file \code{PERSUADE_output.Rmd} is stored within the
package under \verb{inst/rmd/}. Figures are written to a subdirectory \verb{Images/}
inside the output folder, and the knit environment is initialised with the
supplied \code{PERSUADE} object.
}
\examples{
\dontrun{
f_generate_report(PERSUADE)
}

}
\seealso{
\code{\link[=f_PERSUADE]{f_PERSUADE()}}
}
