% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERSUADE_output_functions.R
\name{f_plot_smoothed_hazard}
\alias{f_plot_smoothed_hazard}
\title{Smoothed Hazard Function Plot}
\usage{
f_plot_smoothed_hazard(PERSUADE)
}
\arguments{
\item{PERSUADE}{A PERSUADE object created by \code{\link[=f_PERSUADE]{f_PERSUADE()}}.}
}
\value{
A base R plot of smoothed hazards by group.
}
\description{
Plots smoothed hazard estimates for each group in the PERSUADE object.
}
\examples{
\dontrun{
f_plot_smoothed_hazard(PERSUADE)
}
}
