% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERSUADE_output_functions.R
\name{f_generate_report}
\alias{f_generate_report}
\title{Generate PDF Report for a PERSUADE Analysis}
\usage{
f_generate_report(PERSUADE)
}
\arguments{
\item{PERSUADE}{A PERSUADE object returned by \code{\link[=f_PERSUADE]{f_PERSUADE()}}.}
}
\value{
A length-1 character string giving the absolute path to the generated
PDF, returned invisibly.
}
\description{
Save the PERSUADE object and render a PDF report using \code{PERSUADE_output.Rmd}.
The function creates an output directory \verb{<name>_output}, saves \code{PERSUADE.RData},
and writes \verb{<name>.pdf} into that directory.
}
\details{
The R Markdown file \code{PERSUADE_output.Rmd} must be available in the
working directory. Figures are written to a subdirectory \verb{Images/} and the
knit environment is initialised with the supplied \code{PERSUADE} object.
}
\examples{
\dontrun{
PERSUADE <- f_PERSUADE(...)
pdf_path <- f_generate_report(P)
pdf_path
}
}
\seealso{
\code{\link[=f_PERSUADE]{f_PERSUADE()}}
}
