% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERSUADE_output_functions.R
\name{f_plot_cure_surv_model}
\alias{f_plot_cure_surv_model}
\title{Cure Survival Model Overlay}
\usage{
f_plot_cure_surv_model(PERSUADE, model_index = 1)
}
\arguments{
\item{PERSUADE}{A PERSUADE object created by \code{\link[=f_PERSUADE]{f_PERSUADE()}}.}

\item{model_index}{Integer. Index of the cure model in \code{PERSUADE$surv_pred$model$cure}.}
}
\value{
A base R plot of KM curves with cure model overlays.
}
\description{
Overlays a fitted cure survival model on KM curves, including shaded KM
confidence bands per group.
}
\examples{
\dontrun{
f_plot_cure_surv_model(PERSUADE, model_index = 1)
}

}
\seealso{
\code{\link[=f_PERSUADE]{f_PERSUADE()}}
}
