% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERSUADE_output_functions.R
\name{f_plot_diag_cure_surv_model}
\alias{f_plot_diag_cure_surv_model}
\title{Diagnostic Plot for Cure Survival Models}
\usage{
f_plot_diag_cure_surv_model(PERSUADE, model_index = 1)
}
\arguments{
\item{PERSUADE}{A PERSUADE object created by \code{\link[=f_PERSUADE]{f_PERSUADE()}}.}

\item{model_index}{Integer. Index of the cure model in
\code{PERSUADE$surv_pred$model$cure}.}
}
\value{
A base R diagnostic plot for the selected cure survival model.
}
\description{
Produces diagnostic plots for mixture and non-mixture cure survival models,
using transformations depending on the underlying distribution
(Weibull, log-normal, log-logistic).
}
\examples{
\dontrun{
f_plot_diag_cure_surv_model(PERSUADE, model_index = 1)
}
}
