% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERSUADE_output_functions.R
\name{f_plot_km_survival_base}
\alias{f_plot_km_survival_base}
\title{Plot Kaplan-Meier Survival Curves (Base R)}
\usage{
f_plot_km_survival_base(PERSUADE)
}
\arguments{
\item{PERSUADE}{A PERSUADE object created by \code{\link[=f_PERSUADE]{f_PERSUADE()}}.}
}
\value{
A base R plot.
}
\description{
Generates Kaplan-Meier survival plots from a PERSUADE object using base R graphics,
with shaded confidence intervals and group-specific legends.
}
\examples{
\dontrun{
f_plot_km_survival_base(PERSUADE)
}

}
\seealso{
\code{\link[=f_PERSUADE]{f_PERSUADE()}}
}
